# openapi/openapi.v1.yaml
openapi: 3.1.0
info:
  title: Energy Allocation API
  version: 1.0.0
  description: >-
    Tenant-aware API for onboarding facilities, ingesting data, generating model-driven
    recommendations, and tracking realized savings.
servers:
  - url: https://api.energyallocation.com
security:
  - oauth2: [client_admin, facility_manager, energy_analyst]
paths:
  /v1/clients/{client_id}/facilities:
    get:
      tags: [Onboarding]
      summary: List facilities by client
      operationId: listFacilities
      parameters:
        - $ref: '#/components/parameters/clientId'
        - $ref: '#/components/parameters/status'
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/cursor'
      responses:
        '200':
          description: Facility list
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/Facility' }
                  next_cursor: { type: [string, 'null'] }
                required: [items]

    post:
      tags: [Onboarding]
      summary: Create facility
      operationId: createFacility
      parameters:
        - $ref: '#/components/parameters/clientId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateFacilityRequest'
      responses:
        '201':
          description: Facility created in draft state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateFacilityResponse'
  /v1/facilities/{facility_id}/connectors:
    post:
      tags: [Onboarding]
      summary: Attach connector to facility
      operationId: attachConnector
      parameters:
        - $ref: '#/components/parameters/facilityId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateConnectorRequest'
      responses:
        '201':
          description: Connector created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Connector'
  /v1/connectors/{connector_id}/validate:
    post:
      tags: [Onboarding]
      summary: Validate connector health and channels
      operationId: validateConnector
      parameters:
        - $ref: '#/components/parameters/connectorId'
      responses:
        '200':
          description: Validation complete
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConnectorValidationResult'
  /v1/connectors/{connector_id}/channels:
    get:
      tags: [Ingestion]
      summary: List detected channels
      operationId: listConnectorChannels
      parameters:
        - $ref: '#/components/parameters/connectorId'
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/cursor'
      responses:
        '200':
          description: Paginated channel list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChannelListResponse'
  /v1/facilities/{facility_id}/appliances:
    post:
      tags: [Modeling]
      summary: Create appliance and channel mappings
      operationId: createAppliance
      parameters:
        - $ref: '#/components/parameters/facilityId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateApplianceRequest'
      responses:
        '201':
          description: Appliance created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Appliance'
...
  /v1/facilities/{facility_id}/recommendations:
    get:
      tags: [Recommendations]
      summary: List recommendations
      operationId: listRecommendations
      parameters:
        - $ref: '#/components/parameters/facilityId'
        - $ref: '#/components/parameters/status'
        - $ref: '#/components/parameters/sort'
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/cursor'
      responses:
        '200':
          description: Recommendation list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecommendationListResponse'
  /v1/recommendations/{recommendation_id}/decision:
    post:
      tags: [Recommendations]
      summary: Accept or reject recommendation
      operationId: decideRecommendation
      parameters:
        - $ref: '#/components/parameters/recommendationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecommendationDecisionRequest'
      responses:
        '200':
          description: Decision recorded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Recommendation'
  /v1/recommendations/{recommendation_id}/implementations:
    post:
      tags: [Recommendations]
      summary: Record implementation milestone
      operationId: recordRecommendationImplementation
      parameters:
        - $ref: '#/components/parameters/recommendationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecommendationImplementationRequest'
      responses:
        '201':
          description: Implementation recorded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AsyncJobResponse'
...
components:
  securitySchemes:
    oauth2:
      type: oauth2
      flows:
        clientCredentials:
          tokenUrl: https://auth.energyallocation.com/oauth/token
          scopes:
            client_admin: Manage client-level resources
            facility_manager: Manage facility operations
            energy_analyst: Read analytics and model outputs
            viewer: Read-only access
            ops_admin: Internal operations access
            ml_engineer: Internal ML administration
  parameters:
    clientId: { in: path, name: client_id, required: true, schema: { type: string } }
    facilityId: { in: path, name: facility_id, required: true, schema: { type: string } }
    connectorId: { in: path, name: connector_id, required: true, schema: { type: string } }
    recommendationId: { in: path, name: recommendation_id, required: true, schema: { type: string } }
...
    status: { in: query, name: status, schema: { type: string } }
...
    sort: { in: query, name: sort, schema: { type: string } }
    limit: { in: query, name: limit, schema: { type: integer, default: 50, minimum: 1, maximum: 200 } }
    cursor: { in: query, name: cursor, schema: { type: string } }
...
  schemas:
...
    Facility:
      type: object
      properties:
        id: { type: string }
        client_id: { type: string }
        name: { type: string }
        country: { type: string }
        site_type: { type: string }
        gross_floor_area_m2: { type: number }
        production_profile: { type: string }
        tariff_schedule_id: { type: string }
        timezone: { type: string }
        status: { type: string }
      required: [id, client_id, name, country, site_type, production_profile, timezone, status]
    Appliance:
      type: object
      properties:
        id: { type: string }
        facility_id: { type: string }
        name: { type: string }
        category: { type: string }
        subtype: { type: string }
        rated_power_kw: { type: number }
        control_mode: { type: string }
        zone: { type: string }
        status: { type: string }
      required: [id, facility_id, name, category, rated_power_kw, status]
    Connector:
      type: object
      properties:
        id: { type: string }
        facility_id: { type: string }
        type: { type: string }
        vendor: { type: string }
        source_name: { type: string }
        status: { type: string }
        last_sync_at: { type: string, format: date-time }
      required: [id, facility_id, type, status]
    Channel:
      type: object
      properties:
        channel_id: { type: string }
        label: { type: string }
        unit: { type: string }
        quality_score: { type: number }
        sampling_interval_sec: { type: integer }
      required: [channel_id, label, unit]
    ChannelListResponse:
      type: object
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/Channel' }
        next_cursor: { type: [string, 'null'] }
      required: [items]
    CreateFacilityRequest:
      type: object
      properties:
        name: { type: string }
        country: { type: string }
        site_type: { type: string }
        gross_floor_area_m2: { type: number }
        production_profile: { type: string }
        timezone: { type: string }
      required: [name, country, site_type, production_profile, timezone]
    CreateFacilityResponse:
      type: object
      properties:
        facility_id: { type: string }
        status: { type: string }
        next_step: { type: string }
      required: [facility_id, status, next_step]
    CreateConnectorRequest:
      type: object
      properties:
        type: { type: string }
        vendor: { type: string }
        display_name: { type: string }
        connection:
          type: object
          additionalProperties: true
        credentials_ref: { type: string }
      required: [type, display_name, connection, credentials_ref]
    ConnectorValidationResult:
      type: object
      properties:
        connector_id: { type: string }
        status: { type: string }
        channels_detected: { type: integer }
        sample_window_minutes: { type: integer }
        issues:
          type: array
          items: { type: string }
      required: [connector_id, status, channels_detected, sample_window_minutes, issues]
    CreateApplianceRequest:
      type: object
      properties:
        name: { type: string }
        category: { type: string }
        rated_power_kw: { type: number }
        zone: { type: string }
        channel_mappings:
          type: array
          items:
            type: object
            properties:
              channel_id: { type: string }
              role: { type: string }
            required: [channel_id, role]
      required: [name, category, rated_power_kw, channel_mappings]
...
    Recommendation:
      type: object
      properties:
        recommendation_id: { type: string }
        appliance_id: { type: string }
        title: { type: string }
        projected_annual_savings_eur: { type: number }
        confidence: { type: number }
        effort: { type: string }
        payback_months: { type: number }
        implementation_readiness: { type: number }
        status: { type: string }
      required: [recommendation_id, appliance_id, title, status]
    RecommendationListResponse:
      type: object
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/Recommendation' }
        total_projected_savings_eur: { type: number }
        next_cursor: { type: [string, 'null'] }
      required: [items, total_projected_savings_eur]
    RecommendationDecisionRequest:
      type: object
      properties:
        decision: { type: string, enum: [accepted, rejected] }
        owner_user_id: { type: string }
        target_implementation_date: { type: string, format: date }
        notes: { type: string }
      required: [decision]
    RecommendationImplementationRequest:
      type: object
      properties:
        implementation_date: { type: string, format: date }
        verification_window_days: { type: integer, default: 30 }
        notes: { type: string }
      required: [implementation_date]
...
    AsyncJobResponse:
      type: object
      properties:
        job_id: { type: string }
        status: { type: string, enum: [queued, running, completed, failed] }
      required: [job_id, status]
...